ISSUE #6118 - Updated use of drop pins for the more generic way and coordproperty to avoid re selections and unnecessary remove/adds

+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsForm/properties/coordsProperty/coordsProperty.component.tsx
@@ -125,15 +125,13 @@ export const CoordsProperty = ({ value, label, onChange, onBlur, required, error
 	const refreshPin = () => {
 		if (!isViewer) return;
 
-		if (prevValue.current) {
+		if (prevValue.current && !hasPin) {
 			ViewerService.removePin(pinId);
 		}
 
 		if (hasPin) {
 			ViewerService.showPin(toPin(name, template, minimumPinTicket, false, value));
 		}
-
-		if (isSelected) ViewerService.setSelectionPin({ id: pinId, isSelected });
 	};
 
 	// Update pin when colour or icon changes
